pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-indexer"
version = "8.0.0"
description = "AI Context Indexer – Production-Ready for 2026 LLM Ecosystems"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "pydantic>=2.0",
    "pathspec>=0.11",
]

[project.optional-dependencies]
full = [
    "tree-sitter>=0.22",
    "tree-sitter-python",
    "tree-sitter-typescript",
    "tree-sitter-javascript",
    "tiktoken",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "tomli; python_version < '3.11'",
]
dev = [
    "pytest>=8.0",
    "mypy>=1.9",
    "ruff>=0.4",
]

[project.scripts]
ai-indexer = "ai_indexer.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ai_indexer = ["py.typed"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"
